Space analysis tool for #1004
This commit contains a (slightly hacky, untested) space analysis tool
for investigation into issue #1004, on how to efficiently represent
PageRank results.

When run, it produces various data structures that contain alternate
representations of the data inside a `PagerankNodeDecomposition`, and
saves them to disk in the $SOURCECRED_DIRECTORY. It also prints an
analysis showing the space requirements for different ways of organizing
the pagerank data.

Note that for large repositories, it may be necessary to pass
`--max-old-space-size=8192`

Example Usage:
```
> node bin/sourcecred.js load sourcecred/sourcecred
> node bin/sourcecred.js analyze sourcecred/sourcecred

| Name | Uncompressed Size(M) | Compressed Size(M) | Uncompressed Multiple | Compressed Multiple |
| --- | --- | --- | --- | --- |
pND | 19.1 | 1.0 | 6.41 | 3.54
full objects | 6.7 | 0.5 | 2.25 | 1.96
optimal storage | 3.0 | 0.3 | 1.00 | 1.00

> ls -sh $SOURCECRED_DIRECTORY/data/sourcecred/sourcecred | grep json

3.0M edgeAddressToEdgeWeightObject.json
172K edgeAddressToEdgeWeightObject.json.gz
2.9M graph.json
276K graph.json.gz
640K nodeAddressToScoreObject.json
104K nodeAddressToScoreObject.json.gz
 19M pagerankNodeDecomposition.json
956K pagerankNodeDecomposition.json.gz
 12K unsortedEdgeWeightArray.json
 12K unsortedEdgeWeightArray.json.gz
 12K unsortedNodeScoreArray.json
 12K unsortedNodeScoreArray.json.gz
```