Implement to/fromJSON for PagerankNodeDecomp
For #967, we need to actually serialize the results from running
PageRank so that others can consume it. This commit adds toJSON and
fromJSON methods for PagerankNodeDecomposition so that we can do just
that.

Test plan: Unit tests to verify round trip stability of the JSON have
been added. Run `yarn test`.