Add `WeightedGraph` and `createWeightedGraph`
This commit adds the `WeightedGraph` type to `graphToMarkovChain`. It
has the following type signature:

```
export type WeightedGraph = {|
  +graph: Graph,
  +edgeWeights: Map<EdgeAddressT, EdgeWeight>,
  +nodeTotalOutWeights: Map<NodeAddressT, number>,
  +syntheticLoopWeight: number,
|};
```

The WeightedGraph bundles a graph along with all the metadata needed to
run PageRank on it, and to interpret the PageRank results. The
syntheticLoopWeight and edgeWeights are clearly necessary so that we can
create a weighted markov chain from the graph. The nodeTotalOutWeights
is an optimization suggested by @wchargin that makes it much faster to
compute the normalized weights for those connections. (We could, in
principle, store normalized edge weights directly, but this would
drastically reduce the compressibility of the WeightedGraph).

This commit also adds a tested function, `createWeightedGraph`, which
does exactly what you'd expect from the name. It does not actually
modify any existing code to consume WeightedGraphs—that will come in a
followon commit.

Test plan: Run `yarn test`; unit tests were added for
`createWeightedGraph`.

For more context on why we want a WeightedGraph type, read the
discussion in #1004.

This is part of the #967 work group.