Pass weightedGraph and scores along with pND
Previously, `pagerank` returned just a PagerankNodeDecomposition,
and this was piped down into the shared props of PagerankTable.

As of this commit, `scores` and a `weightedGraph` are returned and piped
along too. This is the first step towards removing the
PagerankNodeDecomposition and replacing it with the scores and
weightedGraph, as discussed in #1004.

Test plan: Unit tests updated.