Add `decomposeNodeScore`
This adds a function for decomposing the score of a single node.
This is equivalent to running the full `decompose` function, but scoped
to a particular node, and (assuming you only want results for that
node), much more performant.

Test plan: Added unit tests to verify that the behavior of
`decomposeNodeScore` is consistent with `decompose` (which is already
tested).