Report Markov Chain convergence statistics
This commit modifies `markovChain.findStationaryDistribution` so that
in addition to returning the final distribution, it also provides
information on how many iterations were required to get that
distribution, and on what the final convergence delta was.

This is motivated by the proposed API for the new PagerankGraph (see
[#1020]). Also, I think it makes a nice addition to the test code.

Test plan:
Unit tests are updated, and `yarn test` passes.

[#1020]: https://github.com/sourcecred/sourcecred/issues/1020

Thanks to @BrianLitwin for semi-pair-programming it